+++ src/main/java/itbank/pethub/service/ShopService.java
@@ -1,0 +1,4 @@
+package itbank.pethub.service;
+
+public class ShopService {
+}
