+++ src/main/resources/templates/header.html
@@ -30,7 +30,7 @@
             <span class="text-lg font-semibold">PetHub</span>
         </a>
         <nav class="hidden md:flex gap-6">
-            <a class="text-gray-500 hover:text-gray-900" href="/order/Items">
+            <a class="text-gray-500 hover:text-gray-900" href="/shop/Items">
                 Shop
             </a>
             <a class="text-gray-500 hover:text-gray-900" href="#">
