--- src/main/resources/templates/order/orderUpdate.html
@@ -1,104 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>주문수정</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            background-color: #f8f9fa;
-            color: #343a40;
-            margin: 0;
-            padding: 20px;
-        }
-
-        h3 {
-            text-align: center;
-            color: #007bff;
-        }
-
-        table {
-            width: 100%;
-            border-collapse: collapse;
-            margin-bottom: 20px;
-            background-color: #ffffff;
-            box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-        }
-
-        th, td {
-            border: 1px solid #dee2e6;
-            padding: 10px;
-            text-align: center;
-        }
-
-        th {
-            background-color: #007bff;
-            color: white;
-        }
-
-        tr:nth-child(even) {
-            background-color: #f2f2f2;
-        }
-
-        input[type="text"], input[type="number"] {
-            width: 100%;
-            padding: 5px;
-            box-sizing: border-box;
-        }
-
-        button {
-            display: block;
-            width: 200px;
-            margin: 20px auto 0;
-            padding: 10px;
-            font-size: 16px;
-            color: white;
-            background-color: #007bff;
-            border: none;
-            border-radius: 5px;
-            cursor: pointer;
-        }
-
-        button:hover {
-            background-color: #0056b3;
-        }
-
-        input[name="delivery_id"] {
-            visibility: hidden;
-            width: 0;
-            height: 0;
-            padding: 0;
-            border: none;
-            margin: 0;
-        }
-    </style>
-</head>
-<body>
-
-<form method="POST">
-    <table class="board">
-        <tr>
-            <th>주문 번호</th>
-            <th>상품 번호</th>
-            <th>단위 가격</th>
-            <th>상품 개수</th>
-            <th>총 가격</th>
-            <th>쿠폰 여부</th>
-            <th>주문 취소</th>
-            <th>주문 삭제</th>
-        </tr>
-
-        <tr th:each="row : ${list}">
-            <td th:text="${row.order_id}"></td>
-            <td th:text="${row.order_item}"></td>
-            <td th:text="${row.order_price}"></td>
-            <td><input name="count" type="number" th:value="${row.count}"></td>
-            <td th:text="${row.order_price * row.count}"></td>
-            <td><input name="coupon_check" type="number" th:value="${row.coupon_check}"></td>
-        </tr>
-    </table>
-
-    <button>주문 수정</button>
-</form>
-</body>
-</html>
