+++ src/main/resources/templates/shop/DetailPage.html
@@ -86,7 +86,7 @@
     </div>
 
     <div>
-        <form th:action="@{'/order/DetailPage/' + ${product.id}}" method="POST">
+        <form th:action="@{'/shop/DetailPage/' + ${product.id}}" method="POST">
             <input type="hidden" name="productId" th:value="${product.id}" />
             <input type="hidden" id="formQuantity" name="quantity" th:value="${1}" />
             <button type="submit">구매</button>
