+++ src/main/resources/templates/shop/Message.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Message Page</title>
+</head>
+<body>
+    <script th:inline="javascript">
+        /*<![CDATA[*/
+        var row = /*[[${row}]]*/ 0;
+        var msg = /*[[${msg}]]*/ '';
+        var path = /*[[${path}]]*/ '';
+
+        if (row !== 0) {
+            alert(msg);
+            location.href = path;
+        } else {
+            alert(msg);
+            history.go(-1);
+        }
+        /*]]>*/
+    </script>
+</body>
+</html>
