+++ src/main/resources/templates/shop/shopHeader.html
